refactor(adk): consolidate attack tests into turn_loop_test.go
Move all 12 TestAttack_* functions from attack_test.go into
turn_loop_test.go and delete the now-empty file. No test logic changes.
